refactored messages

+++ src/main/scala/com/oohish/wire/NetworkParameters.scala
@@ -7,8 +7,6 @@ import com.oohish.peermessages.Tx
 import com.oohish.structures.OutPoint
 import com.oohish.structures.TxIn
 import com.oohish.structures.TxOut
-import com.oohish.structures.int64_t
-import com.oohish.structures.uint32_t
 
 /*
  * Taken from bitcoinj's NetworkParameters
@@ -21,7 +19,7 @@ trait NetworkParameters {
   def packetMagic: Long
   def port: Int
   def proofOfWorkLimit: BigInt
-  def PROTOCOL_VERSION: uint32_t = uint32_t(70001)
+  def PROTOCOL_VERSION: Long = 70001
   def targetTimespan: Int
   def dnsSeeds: List[String]
 
@@ -35,28 +33,28 @@ object MainNetParams extends NetworkParameters {
   def addressHeader: Byte = 0
   def genesisBlock: Block =
     Block(
-      uint32_t(1), //version
+      1, //version
       "0000000000000000000000000000000000000000000000000000000000000000", //prev block
       "4a5e1e4baab89f3a32518a88c31bc87f618f76673e2cc77ab2127b7afdeda33b", //merkle root
-      uint32_t(1231006505), //timestamp
-      uint32_t(486604799), //bits
-      uint32_t(2083236893), //nonce
+      1231006505, //timestamp
+      486604799, //bits
+      2083236893, //nonce
       List(
         Tx(
-          uint32_t(1), //version
+          1, //version
           List(
             TxIn(
               OutPoint(
                 "0000000000000000000000000000000000000000000000000000000000000000", //prev block
-                uint32_t(4294967295L) //n
+                4294967295L //n
                 ),
               "04ffff001d0104455468652054696d65732030332f4a616e2f32303039204368616e63656c6c6f72206f6e206272696e6b206f66207365636f6e64206261696c6f757420666f722062616e6b73", // txins
-              uint32_t(Int.MaxValue.toLong * 2))),
+              Int.MaxValue.toLong * 2)),
           List(
             TxOut(
-              int64_t(5000000000L),
+              5000000000L,
               "04678afdb0fe5548271967f1a67130b7105cd6a828e03909a67962e0ea1f61deb649f6bc3f4cef38c4f35504e51ec112de5c384df7ba0b8d578a4c702b6bf11d5f OP_CHECKSIG")), // txins
-          uint32_t(0) //lockTime
+          0 //lockTime
           )))
   def interval: Int = INTERVAL
   def packetMagic: Long = 0xD9B4BEF9
@@ -76,12 +74,12 @@ object TestNet3Params extends NetworkParameters {
   def addressHeader: Byte = 0
   def genesisBlock: Block =
     Block(
-      uint32_t(1), //version
+      1, //version
       "0000000000000000000000000000000000000000000000000000000000000000", //prev block
       "4a5e1e4baab89f3a32518a88c31bc87f618f76673e2cc77ab2127b7afdeda33b", //merkle root
-      uint32_t(1296688602), //timestamp
-      uint32_t(486604799), //bits
-      uint32_t(414098458), //nonce
+      1296688602, //timestamp
+      486604799, //bits
+      414098458, //nonce
       List())
   def interval: Int = INTERVAL
   def packetMagic: Long = 0x0709110B
