+++ src/test/scala/com/oohish/chain/BlockSpec.scala
@@ -6,7 +6,6 @@ import org.scalatest.FlatSpec
 import org.scalatest.Matchers
 
 import com.oohish.peermessages.Block
-import com.oohish.structures.uint32_t
 import com.oohish.util.HexBytesUtil
 import com.oohish.wire.MainNetParams
 
@@ -16,12 +15,12 @@ class BlockSpec extends FlatSpec with Matchers {
 
     val bh = {
       Block(
-        uint32_t(1), //version
+        1, //version
         "6fe28c0ab6f1b372c1a6a246ae63f74f931e8365e15a089c68d6190000000000", //prev block
         "4a5e1e4baab89f3a32518a88c31bc87f618f76673e2cc77ab2127b7afdeda33b", //merkle root
-        uint32_t(1231006505), //timestamp
-        uint32_t(486604799), //bits
-        uint32_t(2083236893), //nonce
+        1231006505, //timestamp
+        486604799, //bits
+        2083236893, //nonce
         List() //number of transactions
         )
     }
